Add UIView extensions functions to the existing one
